+++ src/core/component/watch/create.ts
@@ -6,11 +6,11 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import watch, { mute, unmute, unwrap, getProxyType } from 'core/object/watch';
+import watch, { mute, unmute, unwrap, getProxyType, isProxy } from 'core/object/watch';
 
 import { getPropertyInfo, PropertyInfo } from 'core/component/reflection';
 import { proxyGetters } from 'core/component/engines';
-import { ComponentInterface,  WatchOptions, RawWatchHandler } from 'core/component/interface';
+import { ComponentInterface, WatchOptions, RawWatchHandler } from 'core/component/interface';
 
 import { tiedWatchers, watcherInitializer, fakeCopyLabel } from 'core/component/watch/const';
 import { cloneWatchValue } from 'core/component/watch/clone';
@@ -22,9 +22,10 @@ import { attachDynamicWatcher } from 'core/component/watch/helpers';
  */
 export function createWatchFn(component: ComponentInterface): ComponentInterface['$watch'] {
 	const
-		watchCache = Object.createDict();
+		watchCache = new Map();
 
-	return (path, optsOrHandler, rawHandler?) => {
+	// eslint-disable-next-line @typescript-eslint/typedef
+	return function watchFn(this: unknown, path, optsOrHandler, rawHandler?) {
 		if (component.isFlyweight) {
 			return null;
 		}
@@ -39,46 +40,93 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 
 		} else {
 			handler = rawHandler;
-			opts = optsOrHandler || {};
+			opts = optsOrHandler ?? {};
+		}
+
+		let
+			info: PropertyInfo;
+
+		if (Object.isString(path)) {
+			info = getPropertyInfo(path, component);
+
+		} else {
+			info = isProxy(path) ? {ctx: path} : path;
+		}
+
+		if (!Object.isString(info.type)) {
+			Object.assign(info, {
+				type: 'mounted',
+				originalPath: info.path,
+				fullPath: info.path
+			});
 		}
 
 		const
-			info: PropertyInfo = Object.isString(path) ? getPropertyInfo(path, component) : path,
-			propCtx = info.ctx.unsafe,
-			ctxParams = propCtx.meta.params;
-
-		const
-			isAccessor = info.type === 'accessor' || info.type === 'computed' || info.accessor,
+			isDefinedPath = Object.size(info.path) > 0,
+			isAccessor = Boolean(info.type === 'accessor' || info.type === 'computed' || info.accessor),
 			watchInfo = isAccessor ? null : proxyGetters[info.type]?.(info.ctx);
 
 		let
 			proxy = watchInfo?.value;
 
-		const
-			needCache = handler.length > 1,
-			ref = info.originalPath;
+		const getVal = () => {
+			switch (info.type) {
+				case 'mounted':
+					return isDefinedPath ? Object.get(info.ctx, info.path) : info.ctx;
 
-		const normalizedOpts = {
+				case 'field':
+					return Object.get(proxy, info.originalPath);
+
+				default:
+					return Object.get(component, info.originalPath);
+			}
+		};
+
+		const wrapDestructor = (destructor) => {
+			if (Object.isFunction(destructor)) {
+				// Every worker that passed to async have a counter with number of consumers of this worker,
+				// but in this case this behaviour is redundant and can produce an error,
+				// that why we wrap original destructor with a new function
+				component.unsafe.$async.worker(() => destructor());
+			}
+
+			return destructor;
+		};
+
+		const normalizedOpts = <WatchOptions>{
 			collapse: true,
 			...opts,
 			...watchInfo?.opts
 		};
 
+		const
+			needCache = handler.length > 1 && (isDefinedPath || normalizedOpts.collapse !== false),
+			originalHandler = handler;
+
 		let
 			oldVal;
 
-		const
-			originalHandler = handler,
-			getVal = () => Object.get(info.type === 'field' ? proxy : component, info.originalPath);
-
 		if (needCache) {
-			oldVal = watchCache[ref] = ref in watchCache ?
-				watchCache[ref] :
-				opts.immediate || !isAccessor ? cloneWatchValue(getVal(), opts) : undefined;
+			let
+				cacheKey;
 
-			handler = (val, _, ...args) => {
+			if (Object.isString(info.originalPath)) {
+				cacheKey = [info.originalPath];
+
+			} else {
+				cacheKey = Array.concat([info.ctx], info.path);
+			}
+
+			if (Object.has(watchCache, cacheKey)) {
+				oldVal = Object.get(watchCache, cacheKey);
+
+			} else {
+				oldVal = normalizedOpts.immediate || !isAccessor ? cloneWatchValue(getVal(), normalizedOpts) : undefined;
+				Object.set(watchCache, cacheKey, oldVal);
+			}
+
+			handler = (val, _, i) => {
 				if (isAccessor) {
-					// tslint:disable-next-line:prefer-conditional-expression
 					if (normalizedOpts.collapse) {
 						val = Object.get(info.ctx, info.accessor ?? info.name);
 
@@ -87,14 +135,22 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 					}
 				}
 
-				const res = originalHandler.call(this, val, oldVal, ...args);
-				oldVal = watchCache[ref] = cloneWatchValue(val, opts);
+				if (!isDefinedPath && Object.isArray(i?.path)) {
+					oldVal = Object.get(oldVal, [i.path[0]]);
+				}
+
+				const
+					res = originalHandler.call(this, val, oldVal, i);
+
+				oldVal = cloneWatchValue(isDefinedPath ? val : getVal(), normalizedOpts);
+				Object.set(watchCache, cacheKey, oldVal);
+
 				return res;
 			};
 
 			handler[tiedWatchers] = originalHandler[tiedWatchers];
 
-			if (opts.immediate) {
+			if (normalizedOpts.immediate) {
 				const val = oldVal;
 				oldVal = undefined;
 				handler.call(component, val);
@@ -102,8 +158,7 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 
 		} else {
 			if (isAccessor) {
-				handler = (val, oldVal, ...args) => {
-					// tslint:disable-next-line:prefer-conditional-expression
+				handler = (val, _, i) => {
 					if (normalizedOpts.collapse) {
 						val = Object.get(info.ctx, info.accessor ?? info.name);
 
@@ -111,21 +166,35 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 						val = Object.get(component, info.originalPath);
 					}
 
-					return originalHandler.call(this, val, oldVal, ...args);
+					if (!isDefinedPath && Object.isArray(i?.path)) {
+						oldVal = Object.get(oldVal, [i.path[0]]);
+					}
+
+					const res = originalHandler.call(this, val, oldVal, i);
+					oldVal = isDefinedPath ? val : getVal();
+
+					return res;
 				};
 			}
 
-			if (opts.immediate) {
+			if (normalizedOpts.immediate) {
 				handler.call(component, getVal());
 			}
 		}
 
-		const rootOrFunctional = Boolean(
-			ctxParams.root || ctxParams.functional === true
-		);
+		let
+			rootOrFunctional = false;
 
-		if (proxy) {
-			if (!watchInfo) {
+		if (info.type !== 'mounted') {
+			const
+				propCtx = info.ctx.unsafe,
+				ctxParams = propCtx.meta.params;
+
+			rootOrFunctional = Boolean(ctxParams.root) || ctxParams.functional === true;
+		}
+
+		if (proxy != null) {
+			if (watchInfo == null) {
 				return null;
 			}
 
@@ -138,6 +207,9 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 						mute(proxy);
 						proxy[info.name] = info.ctx[info.name];
 						unmute(proxy);
+
+						const
+							propCtx = info.ctx.unsafe;
 
 						Object.defineProperty(info.ctx, info.name, {
 							enumerable: true,
@@ -177,10 +249,11 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 						});
 
 					} else {
+						// eslint-disable-next-line @typescript-eslint/unbound-method
 						unwatch = watch(proxy, info.path, normalizedOpts, handler).unwatch;
 					}
 
-					return unwatch;
+					return wrapDestructor(unwatch);
 				}
 
 				case 'prop': {
@@ -200,9 +273,10 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 							destructors.pop();
 						}
 
+						// eslint-disable-next-line no-use-before-define
 						attachDeepProxy();
 
-						if (value?.[fakeCopyLabel]) {
+						if (value?.[fakeCopyLabel] === true) {
 							return;
 						}
 
@@ -226,7 +300,7 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 							const
 								tiedLinks = handler[tiedWatchers];
 
-							if (tiedLinks) {
+							if (Object.isArray(tiedLinks)) {
 								for (let i = 0; i < tiedLinks.length; i++) {
 									const modifiedInfo = {
 										...info,
@@ -243,7 +317,8 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 						});
 
 					} else {
-						unwatch = watch(proxy, info.path, normalizedOpts, watchHandler).unwatch;
+						// eslint-disable-next-line @typescript-eslint/unbound-method
+						unwatch = watch(proxy, info.path, <any>normalizedOpts, watchHandler).unwatch;
 					}
 
 					destructors.push(unwatch);
@@ -255,7 +330,7 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 						const
 							proxyVal = Object.get(unwrap(proxy), info.path);
 
-						if (getProxyType(proxyVal)) {
+						if (getProxyType(proxyVal) != null) {
 							const normalizedOpts = {
 								collapse: true,
 								...opts,
@@ -275,6 +350,7 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 								}
 							};
 
+							// eslint-disable-next-line @typescript-eslint/unbound-method
 							const {unwatch} = watch(<object>proxyVal, normalizedOpts, watchHandler);
 							destructors.push(unwatch);
 						}
@@ -282,16 +358,23 @@ export function createWatchFn(component: ComponentInterface): ComponentInterface
 
 					attachDeepProxy();
 
-					return () => {
+					return wrapDestructor(() => {
 						for (let i = 0; i < destructors.length; i++) {
 							destructors[i]();
 						}
-					};
+					});
 				}
+
+				default:
+					// Loopback
 			}
 
-			const {unwatch} = watch(proxy, info.path, normalizedOpts, handler);
-			return unwatch;
+			// eslint-disable-next-line @typescript-eslint/unbound-method
+			const {unwatch} = isDefinedPath ?
+				watch(proxy, info.path, normalizedOpts, handler) :
+				watch(proxy, normalizedOpts, handler);
+
+			return wrapDestructor(unwatch);
 		}
 
 		return attachDynamicWatcher(component, info, opts, handler);