+++ src/core/component/decorators/CHANGELOG.md
@@ -11,6 +11,23 @@ Changelog
 
 ## v3.0.0-rc.37 (2020-07-20)
 
+#### :boom: Breaking Change
+
+* Now all accessors with dependencies are cacheable by default
+
+#### :rocket: New Feature
+
+* Added support of mounted watchers
+
+```js
+class bFoo {
+  @computed({watchable: true})
+  get remoteState(): typeof anotherWatcher {
+    return anotherWatcher;
+  }
+}
+```
+
 #### :house: Internal
 
 * Fixed ESLint warnings