+++ src/core/component/reflection/CHANGELOG.md
@@ -9,12 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.0.0-rc.36 (2020-07-13)
+## v3.0.0-rc.37 (2020-07-20)
 
 #### :rocket: New Feature
 
-* Added the `worker` option
+* Added support for mounted properties
 
 #### :house: Internal
 
-* Refactoring
+* Fixed ESLint warnings